pyproject.toml
[project]
name = "empresas-certificadas-etl"
version = "0.1.0"
description = "ETL pipeline that consolidates certified tourism companies datasets"
authors = [{name = "UACh Data Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.1.14"
]

[tool.ruff]
line-length = 100
select = ["E", "F", "B", "I"]
